+++ src/components/form/InputField/inputField.stories.js
@@ -139,3 +139,30 @@ WithCharacterLimit.args = {
   max_length_suffix: 'characters remaining.',
   max_length: 40,
 };
+export const Disabled = Template.bind({});
+Disabled.args = {
+  id: 'email',
+  name: 'email',
+  label: 'Email or username',
+  caption: 'Enter your email address or username.',
+  required: '',
+  error: false,
+  error_message: 'Unrecognized username or password',
+  value: 'test@example.com',
+  type: 'email',
+  disabled : true,
+};
+
+export const ReadOnly = Template.bind({});
+ReadOnly.args = {
+  id: 'email',
+  name: 'email',
+  label: 'Email or username',
+  caption: 'Enter your email address or username.',
+  required: '',
+  error: false,
+  error_message: 'Unrecognized username or password',
+  value: 'test@example.com',
+  type: 'email',
+  read_only : true,
+};